# Fix a bus congestion when polling the chip too often.