--- a/probe-rs/src/architecture/arm/core/m4.rs
+++ b/probe-rs/src/architecture/arm/core/m4.rs
@@ -352,7 +352,6 @@ impl<'probe> CoreInterface for M4<'probe> {
         // Wait until halted state is active again.
         let start = Instant::now();
 
-        let mut n = 0;
         while start.elapsed() < timeout {
             let dhcsr_val = Dhcsr(self.memory.read_word_32(Dhcsr::ADDRESS)?);
             if dhcsr_val.s_halt() {
@@ -361,8 +360,7 @@ impl<'probe> CoreInterface for M4<'probe> {
 
                 return Ok(());
             }
-            std::thread::sleep(Duration::from_millis(1 << n));
-            n += 1;
+            std::thread::sleep(Duration::from_millis(1));
         }
         Err(Error::Probe(DebugProbeError::Timeout))
     }