Previously, when waiting for a flash algorithm routine to complete, the algorithm on the host side would poll the core state relentlessly, leading to frequent timeouts when polling longer operations, such as sector erase, for completion. Now we do two things: a) we honor the timeouts given to us with the flash algorithm description files properly, and b) we poll the core at maximum once per 10ms with increasing delays each time. This results in significantly less traffic and as a nice byproduct way less log traffic.