@Tiwalun 2021-05-09
I think it's a good idea to poll less frequently. I'm not sure if increasing the wait time in this way is a good idea,  I see a bit of a conflict with the timeout. For example, when a 100 ms timeout is specified, with this PR we poll at the following times:

- 0 ms
- 1 ms
- 3 ms
- 7 ms
- 15 ms
- 31 ms
- 63 ms

The next time we would poll would be after 127 ms, but then we run into the timeout. In the end, we effectively we have a 63 ms timeout, not a 100 ms timeout. Maybe we should poll once again at the end, so that the full timeout is used?