@Tiwalun 2021-05-09
And from a quick benchmark, this PR increases the time needed for flashing on a STM32WB55 discovery board from about 8 s to 8.5 s.  

  &nbsp; | erase | program | total
--- | --- | --- | ---
master | 114 KiB/s | 23 KiB/s  | 8s
PR | 84 KiB/s | 21 KiB/s | 8.5s